+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [QCheck-STM] Add domain bug report generation
+  [\#346](Https://github.com/ocaml-gospel/ortac/pull/346)
 - [QCheck-STM] Make `ortac_show_cmd` look at the models
   [\#353](https://github.com/ocaml-gospel/ortac/pull/353)
 - [QCheck-STM] Refactor bug report printing